Clean up of VS2015 warnings - Part 1

+++ native/ax/MpegDeMux-3/DeMuxFilter.cpp
@@ -985,7 +985,7 @@ HRESULT CDeMuxFilter::CompleteConnect(IPin *pReceivePin, CBasePin* pFilterPin )
 
 BOOL CDeMuxFilter::IsDeclineFilter( char *pFilterName )
 {
-	int len = strlen( pFilterName );
+	size_t len = strlen( pFilterName );
 	char* p = pFilterName;
 
 	while (*p) {*p=tolower(*p);p++;} //incase sensitive
@@ -2059,7 +2059,7 @@ void CDeMuxFilter::dump_data( char* pData, int dwBytes )
 void CDeMuxFilter::dump_sample( IMediaSample *pSample )
 {
 	BYTE* pBuffer;
-	int   Length;
+	size_t   Length;
 	if ( fd == NULL ) return;
 	if ( FAILED(  pSample->GetPointer( &pBuffer ) ) ) return;
 	Length = pSample->GetActualDataLength( );
@@ -2075,7 +2075,7 @@ void CDeMuxFilter::open_dump( )
 {
 	char fname[MAX_PATH];
 	char path[MAX_PATH]={0};
-	int len;
+	size_t len;
 	sprintf_s( fname, sizeof(fname), "DUMP_FILTER_DATA.ENABLE" );
 	fd = fopen( fname, "r" );
 	if ( fd == NULL ) return;
